Re #112 Implement systemtest for fake EPICS client
